% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/holiday_scrape.R
\name{add_black_fri}
\alias{add_black_fri}
\title{Add Black Friday to a calendar year of holidays}
\usage{
add_black_fri(x)
}
\arguments{
\item{x}{A \code{data.frame} for a single year of holidays; internal to \code{holiday_scrape()}.}
}
\description{
Adds Black Friday to a calendar year of holidays
}

