% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adj_base_forecasts.R
\name{adj_base_forecasts}
\alias{adj_base_forecasts}
\title{Adjust baseline forecasts for weekday seasonality.}
\usage{
adj_base_forecasts(baseline, calls, camp_tot, seasonal_wks = 4,
  seasonal_adj_type = c("stl", "ets", "wk_avg", "ensemble"), call_hist,
  beg_year = 2014, end_year = year(Sys.Date()))
}
\arguments{
\item{baseline}{A \code{data.frame} of aggregate forecasts. One of the return
values from \code{create_baseline_forecasts()}.}

\item{calls}{A \code{data.frame} of call data from Phoenix with categorical data. Returned
from \code{data_pull()}.}

\item{camp_tot}{A \code{data.frame} of campaign data from c2g.campaign_response. Should be
combined from the returns of \code{get_model_data()$camp_complete} and
\code{get_model_data()$camp_outstanding}.}

\item{seasonal_wks}{An integer for the number of trailing weeks used to calculate the weekly
seasonal pattern in the response trend. Used for model tuning. Defaults to 4 (for now)}

\item{seasonal_adj_type}{\code{character}. A method for weekday seasonal adjustment.}

\item{call_hist}{A \code{data.frame} of historical calls data on which to examine holiday patterns.}

\item{beg_year}{An integer specifying first year of holidays to pull. Defaults to 2014.}

\item{end_year}{An integer specifying last year of holidays to pull. Defaults to \code{year(Sys.Date())}.}
}
\value{
A \code{data.frame} of call projections by category for the month.
}
\description{
Create exponential smoothing estimates of weekday seasonality
via STL. Use seasonality estimates to adjust baseline forecasts.
}

