% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/holiday_adj.R
\name{apply_hol_adj}
\alias{apply_hol_adj}
\title{Apply holiday adjustments}
\usage{
apply_hol_adj(cur_forecasts, pred_adj)
}
\arguments{
\item{cur_forecasts}{A \code{data.frame}. The internal object \code{projections}.}

\item{pred_adj}{A \code{data.frame}. The internal object \code{pred}.}
}
\description{
This function should only be called internally.
It applies the holiday adjustments to the forecasts
}

