% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/camp_out_calc_adj.R
\name{camp_out_calc_adj}
\alias{camp_out_calc_adj}
\title{Calculate adjustment rate (ie lift) for oustanding campaigns}
\usage{
camp_out_calc_adj(camp_out, channel = "c2g")
}
\arguments{
\item{camp_out}{A \code{data.frame} with data on oustanding campaigns. Is an output
from \code{get_model_data()}.}

\item{channel}{A character string corresponding to the appropriate ODBC connection. Defaults to "c2g"}
}
\value{
A \code{list} of length two: (1) \code{list} of top performing campaigns by class,
(2) the adjustment rate (lift) for each class.
}
\description{
Calculates adjustment rate (ie-lift) for ongoing campaigns by class and
IDs top ongoing capaigns where statistically significant. Class is defined as
\code{interaction(campaign_type, class_of_mail)}.
}

