% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_baseline_forecasts.R
\name{create_baseline_forecasts}
\alias{create_baseline_forecasts}
\title{Create baseline call projection forecasts}
\usage{
create_baseline_forecasts(camp_proj, camp_comp_stats, top_outstanding)
}
\arguments{
\item{camp_proj}{A \code{data.table}. Should be one of the returned values from
\code{get_model_data()}.}

\item{camp_comp_stats}{A \code{data.table} of summary statistics of completed
campaigns computed at the \code{group_by(campaign_type, class_of_mail, days_to_response)}
level.}

\item{top_outstanding}{A \code{list}. Should be the results from \code{camp_out_calc_adj()}.}
}
\value{
A \code{list} of length two with: (1) Daily forecasts at the campaign level; and (2)
aggregated daily forecasts
}
\description{
Creates the baseline call projections for the model based on historical
response rates for each campaign class (\code{campaign_type:class_of_mail}).
}

