% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/camp_out_calc_adj.R
\name{extract_top}
\alias{extract_top}
\title{Extract top / not top campaign data}
\usage{
extract_top(campaigns, rr_stats, channel = "c2g")
}
\arguments{
\item{campaigns}{A \code{data.frame} with data on complete campaigns from [campaign_tracker].}

\item{rr_stats}{A \code{data.frame} with summary statistics on complete campaigns.}

\item{channel}{A character string corresponding to the appropriate ODBC connection. Defaults to "c2g"}
}
\value{
a \code{list} of two \code{data.table}s: (1) top campaign data; (2) not-top campaign data
}
\description{
Pulls campaign data from [campaign_response] for top campaigns and not-top campaigns
by class (\code{campaign_type:class_of_mail})
}

