% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Clean Model data}
\usage{
get_model_data(called_data, channel = "c2g")
}
\arguments{
\item{called_data}{A \code{data.frame} with call data from \code{data_pull()}.}

\item{channel}{A character string corresponding to the appropriate ODBC connection. Defaults to "c2g"}
}
\value{
\code{list} of four \code{data.frame}s: (1) Aggregated call-by-day data, (2) completed
campaign data, (3) oustanding campaign data, (4) all future response days for upcoming campaigns.
}
\description{
Input IB call data, pull campaign response data. Do aggregations, data cleaning,
and find all future dates for oustanding and future campaigns where we expect responses within
the upcoming month.
}

