% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/holiday_adj.R
\name{holiday_adj}
\alias{holiday_adj}
\title{Define and apply holiday adjustments to forecasts}
\usage{
holiday_adj(cur_forecasts, beg_year = 2014, end_year = year(Sys.Date()),
  call_hist, future = FALSE)
}
\arguments{
\item{cur_forecasts}{A \code{data.frame} of the current, unadjusted (for holidays), forecasts.}

\item{beg_year}{An \code{integer} specifying the first year in which you would like holidays}

\item{end_year}{An \code{integer} specifying the last year in which you would like holidays.
Defaults to the current year.}

\item{call_hist}{A \code{data.frame} of historical calls data on which to examine holiday patterns.}

\item{future}{Logical. If \code{TRUE} current month is increased by 1.}
}
\value{
A \code{data.frame} of forecasts that have been adjusted for holidays.
}
\description{
Adjusts current forecasts and adjusts for any holidays in the forecast period.
Appropriate holiday adjustments are found via negative binomial regression.
}

