% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/holiday_scrape.R
\name{holiday_scrape}
\alias{holiday_scrape}
\title{Scrape Holidays from Web}
\usage{
holiday_scrape(beg_year, end_year)
}
\arguments{
\item{beg_year}{An \code{integer} specifying the first year in which you would like holidays}

\item{end_year}{An \code{integer} specifying the last year in which you would like holidays}
}
\value{
A \code{data.frame} of holidays
}
\description{
Scrapes holidays from \url{http://www.timeanddate.com/holidays/us/} for
specified begining and end dates (specified as years)
}

