% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/camp_out_calc_adj.R
\name{id_top_inc}
\alias{id_top_inc}
\title{Identify top performing ongoing campaigns}
\usage{
id_top_inc(complete_camp_rr, top_camp_list, inc_campaigns)
}
\arguments{
\item{complete_camp_rr}{A \code{data.table}. Should be the "comp_rr" internal from
camp_out_calc_adj().}

\item{top_camp_list}{A \code{list} with summary statistics for top performing complete
campaigns. Should be the "camp_sum_stats[[1]]" internal from camp_out_calc_adj().}

\item{inc_campaigns}{A \code{data.table}. Should be the "inc_camp_track" internal
from camp_out_calc_adj().}
}
\description{
Using historical data on complete campaigns and data on incomplete /
ongoing campaigns, identify ongoing campaigns which are expected to be top performing
campaigns.
}

