% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute_miss_rr.R
\name{impute_zero_resp}
\alias{impute_zero_resp}
\title{Impute Zero Response Rate - missing data}
\usage{
impute_zero_resp(dat, days_tracking = 90)
}
\arguments{
\item{dat}{A \code{data.frame} of a completed campaign.}

\item{days_tracking}{An integer corresponding to the days of tracking across campaigns.
Defaults to 90.}
}
\value{
A \code{data.frame} with imputed zero responses on days without responses.
}
\description{
Imputes missing response rate data for single campaign without 90 days of
reported tracking (ie-- days where 0 individuals responded)
}

