% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute_miss_rr.R
\name{impute_zero_resp_all}
\alias{impute_zero_resp_all}
\title{Impute Zero Response Rate - missing data (all campaigns)}
\usage{
impute_zero_resp_all(dat, days_tracking = 90, group = "cell_code",
  comp_camp = TRUE)
}
\arguments{
\item{dat}{A \code{data.frame} of all completed campaigns.}

\item{days_tracking}{An integer corresponding to the days of tracking across campaigns.
Defaults to 90.}

\item{group}{A character vector of length 1 corresponding to the grouping variable.
Defaults to \code{"cell_code"}.}

\item{comp_camp}{\code{Logical}. Are the campaigns complete or not? Defaults to TRUE.}
}
\value{
A \code{data.frame} with imputed zero responses on days without responses.
}
\description{
Wrapper for \code{impute_zero_resp}. Imputes missing response rate data
for all campaigns without 90 days of reported tracking (ie-- days where 0 individuals
responded) in a provided dataset.
}

