% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/camp_out_calc_adj.R
\name{listwise.t.test}
\alias{listwise.t.test}
\title{Listwise t-test}
\usage{
listwise.t.test(df1, df2, alpha = 0.05)
}
\arguments{
\item{df1}{A \code{data.frame} with four columns: (1) a key, (2) sample size,
(3) a vector of means, (4) a vector of SDs}

\item{df2}{A \code{data.frame} with four columns: (1) a key, (2) sample size,
(3) a vector of means, (4) a vector of SDs}

\item{alpha}{The significance level of test. Defaults to 0.05}
}
\description{
Computes pairs of Weltch t-tests for matching
}

