% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_campaign_proj.R
\name{new_campaign_proj}
\alias{new_campaign_proj}
\title{Create backbone for future forecasts}
\usage{
new_campaign_proj(cur_date, future_date, cur_campaigns)
}
\arguments{
\item{cur_date}{\code{Numeric}. Coerced date corresponding to the last day of the current month.}

\item{future_date}{\code{Numeric}. Corresponds to first day after max response date to date.}

\item{cur_campaigns}{\code{data.frame}. Defined in \code{get_model_data()}.}
}
\value{
A \code{data.frame} with all future dates where we expect responses for
each current campaign for use in projection.
}
\description{
Internal function called by \code{clean_model_data()}. Calculates all future dates
where we expect responses for each current campaign for use in projection.
}

