% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/holiday_scrape.R
\name{num_wks}
\alias{num_wks}
\title{Number weeks of a vector of days}
\usage{
num_wks(days, sun_indices)
}
\arguments{
\item{days}{A numeric \code{vector} of weekdays with elements in $[1,7]$ ; Sunday == 1, Sat == 7}

\item{sun_indices}{A numeric \code{vector} of indices denoting which elements of \code{days}
correspond to Sundays.}
}
\description{
Provided a vector of weekdays and a vector of indices corresponding to Sundays,
return a vector corresponding to the week numbers for the days vector
}

