% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_pull.R
\name{pull_call_data}
\alias{pull_call_data}
\title{Pull call data}
\usage{
pull_call_data(channel = "c2g", call_date = "1/01/2014")
}
\arguments{
\item{channel}{A character string corresponding to the appropriate ODBC connection. Defaults to "c2g"}

\item{call_date}{A character string specifying subset \code{where call_date >= date}. Defaults to '1/01/2014'}
}
\value{
a \code{data.frame} with call data
}
\description{
Pull in inbound five9id call data from the appropriate databases / daily dump
}

