% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adj_base_forecasts.R
\name{adj_base_forecasts}
\alias{adj_base_forecasts}
\title{Adjust baseline forecasts for weekday seasonality.}
\usage{
adj_base_forecasts(baseline, calls, camp_tot, seasonal_wks = 4,
  seasonal_adj_type = c("stl", "ets", "wk_avg", "ensemble", "wk_stl"),
  call_hist, beg_year = 2014, end_year = year(Sys.Date()),
  etsmodel = "ZZZ", windsor.q = c(0.25, 0.75), future = FALSE)
}
\arguments{
\item{baseline}{A \code{data.frame} of aggregate forecasts. One of the return
values from \code{create_baseline_forecasts()}.}

\item{calls}{A \code{data.frame} of call data from Phoenix with categorical data. Returned
from \code{data_pull()}.}

\item{camp_tot}{A \code{data.frame} of campaign resposne data from c2g.campaign_response.
Should be combined from the returns of \code{get_model_data()$camp_complete} and
\code{get_model_data()$camp_outstanding}. Used for adjusting calls to c2g tracked reponses.}

\item{seasonal_wks}{An integer for the number of trailing weeks used to calculate the weekly
seasonal pattern in the response trend. Used for model tuning. Defaults to 4 (for now)}

\item{seasonal_adj_type}{\code{character}. A method for weekday seasonal adjustment.}

\item{call_hist}{A \code{data.frame} of historical calls data on which to examine holiday patterns.}

\item{beg_year}{An integer specifying first year of holidays to pull. Defaults to 2014.}

\item{end_year}{An integer specifying last year of holidays to pull. Defaults to \code{year(Sys.Date())}.}

\item{etsmodel}{A three-character string identifying method using the framework terminology
of Hyndman et al. (2002) and Hyndman et al. (2008). The first letter denotes the error type ("A", "M"
or "Z"); the second letter denotes the trend type ("N","A","M" or "Z"); and the third letter denotes
the season type ("N","A","M" or "Z"). In all cases, "N"=none, "A"=additive, "M"=multiplicative and
"Z"=automatically selected.}

\item{windsor.q}{A numeric vector of length 2 in (0,1) for windsorizing forecasts.}

\item{future}{Logical. Passed to \code{holiday_adj()}.}
}
\value{
A \code{data.frame} of call projections by category for the month.
}
\description{
Create exponential smoothing estimates of weekday seasonality
via STL. Use seasonality estimates to adjust baseline forecasts.
}

