% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Clean Model data}
\usage{
get_model_data(channel = "c2g", historical = FALSE, hist_yr, hist_mo)
}
\arguments{
\item{channel}{A character string corresponding to the appropriate ODBC connection. Defaults to "c2g"}

\item{historical}{Logical. Do you want to forecast historical projections for testing?
Defaults to FALSE}

\item{hist_yr}{An integer corresponding to a historical year. Must be $>= 2012$.}

\item{hist_mo}{An integer in $[1,12]$ corresponding to a historical month.}
}
\value{
\code{list} of three \code{data.frame}s: (21 completed
campaign data, (2) oustanding campaign data, (3) all future response days for upcoming campaigns.
}
\description{
Pull campaign response data. Do aggregations, data cleaning,
and find all future dates for oustanding and future campaigns where we expect responses within
the upcoming month.
}

